Fix: 1.8 & 1.9 OptiFine Install
